% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expression_and_activity_matrix.R
\name{get_expression_and_activity_matrix}
\alias{get_expression_and_activity_matrix}
\title{Obtain Expression and Gene Activity Matrices for GRN Construction}
\usage{
get_expression_and_activity_matrix(
  TGs = results_identify_TGs$TGs,
  TFs = TFs_in_JASPAR2024,
  scRNAseq = scRNAseq,
  scATACseq = scATACseq,
  promoter_range = 50,
  path
)
}
\arguments{
\item{TGs}{Character vector of target gene symbols to include in the analysis.
Default is results_identify_TGs$TGs. results_identify_TGs is the output of function identify_TGs.}

\item{TFs}{Character vector of transcription factor symbols to include in the analysis.
Default is TFs_in_JASPAR2024. TFs_in_JASPAR2024 is the output of function get_TGs_from_JASPAR2024.}

\item{scRNAseq}{See in ?identify_TGs.}

\item{scATACseq}{See in ?identify_TGs.}

\item{promoter_range}{Numeric. The genomic range (in base pairs) to extend around transcription start sites for calculating gene activity scores.
Must be set prior to code execution and remain unchanged during runtime.}

\item{path}{Absolute path to the fragment tbi. file (both .tsv.gz and .tbi index must exist in the same directory).
For example, "/mnt/d/pbmc-equation/pbmc_granulocyte_sorted_10k_atac_fragments.tsv.gz".}
}
\value{
A list containing five elements:
\itemize{
\item \code{scRNAseq_for_GRN}: Subsetted scRNA-seq data containing only TGs and TFs
\item \code{scATACseq_for_GRN}: Subsetted scATAC-seq data with gene activity matrix (only TGs and TFs)
\item \code{gene_for_GRN}: charactor vector of TGs and TFs
\item \code{TGs}: Filtered target genes present in both datasets
\item \code{TFs}: Filtered transcription factors present in both datasets
}
}
\description{
This function processes single-cell multi-omics data to extract expression matrices for target genes (TGs) and transcription factors (TFs),
and computes gene activity matrices from scATAC-seq data for GRN (Gene Regulatory Network) construction.
}
\details{
The function performs the following key operations:
\enumerate{
\item Creates a dedicated output directory for organization
\item Extracts expression matrices for TGs and TFs from scRNA-seq data
\item Annotates chromatin fragments and computes gene activity scores from scATAC-seq data
\item Integrates both datasets and filters genes to ensure consistency
\item Returns processed data ready for GRN inference
}
}
\note{
Important requirements:
\itemize{
\item Requires both fragment file and its index (.tbi) in the specified path
\item Uses EnsDb.Hsapiens.v86 for genomic annotations (modify for other species)
\item Temporarily changes working directory but restores upon completion
\item activity is calculated based on ATAC-seq fragment counts in gene bodies and promoter regions
\item This function must run in Linux, as \code{GenomeInfoDb::seqlevelsStyle(annotations) = 'UCSC'} is guaranteed to work in Linux but has very low success rate on Windows.
}
}
\examples{
\dontrun{
# Run in linux.
library(EnsDb.Hsapiens.v86)
results_identify_TGs = base::readRDS("./1.1 Data Preprocessing - Identify TGs By Annotation/results_identify_TGs.rds")
TFs_in_JASPAR2024 = base::readRDS("./1.2 Data Preprocessing - Get TGs From JASPAR2024/TFs_in_JASPAR2024.rds")
scRNAseq = base::readRDS("./scRNAseq.rds")
scATACseq = base::readRDS("./scATACseq.rds")
promoter_range = 50
Basic_Info = get_expression_and_activity_matrix(
  TGs = results_identify_TGs$TGs,
  TFs = TFs_in_JASPAR2024,
  scRNAseq = scRNAseq,
  scATACseq = scATACseq,
  promoter_range = promoter_range,
  path = "/mnt/d/pbmc-equation/pbmc_granulocyte_sorted_10k_atac_fragments.tsv.gz"
)
base::saveRDS(Basic_Info, file = "./1.3 Data Preprocessing - Get Expression And Activity Matrix/Basic_Info.rds")
}
}
