% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TGs_from_JASPAR2024.R
\name{get_TGs_from_JASPAR2024}
\alias{get_TGs_from_JASPAR2024}
\title{Retrieve Transcription Factors Genes (TFs, TFGs) from JASPAR2024 Database}
\usage{
get_TGs_from_JASPAR2024(
  scRNAseq = scRNAseq,
  species = "Homo sapiens",
  collection = c("CORE", "CNE", "PHYLOFACTS", "SPLICE", "POLII", "FAM", "PBM",
    "PBM_HOMEO", "PBM_HLH", "UNVALIDATED")
)
}
\arguments{
\item{scRNAseq}{See in ?identify_TGs.}

\item{species}{Character. Species to query in JASPAR2024 database. Default is "Homo sapiens".
Parameter species and genome must be compatible.
Execute this command to verify: \code{rbioapi::rba_jaspar_species(2024)$species}.}

\item{collection}{Collection Character vector. Specifies which JASPAR sub-collections to search through.
Available options include: "CORE", "CNE", "PHYLOFACTS", "SPLICE", "POLII", "FAM", "PBM", "PBM_HOMEO", "PBM_HLH", "UNVALIDATED".
Default includes all.}
}
\value{
A character vector of transcription factor names that exist in both JASPAR2024 database and the provided scRNAseq data.

#' @details
The function performs the following operations:
\enumerate{
\item Connects to JASPAR2024 database and retrieves position weight matrices (PWMs)
\item Filters TFs to retain only those present in the scRNA-seq data
\item Creates a dedicated directory for output organization
\item Returns a list of validated TF names
}
}
\description{
This function extracts transcription factors Genes (TFs, TFGs) from the JASPAR2024 database that are present in the provided single-cell RNA sequencing data.
It filters TFs based on species and collection parameters, and returns a vector of TF names available in both JASPAR2024 and the input data.
}
\note{
This function temporarily changes the working directory to save results but reverts to the original directory upon completion.
The function assumes that all TFs have potential regulatory relationships with each target gene.
}
\examples{
\dontrun{
# Retrieve human TFs from core JASPAR collections
scRNAseq = base::readRDS("./scRNAseq.rds")
TFs_in_JASPAR2024 = get_TGs_from_JASPAR2024(
  scRNAseq = scRNAseq,
  species = "Homo sapiens",
  collection = "CORE"
)
base::saveRDS(TFs_in_JASPAR2024, file = "./1.2 Data Preprocessing - Get TGs From JASPAR2024/TFs_in_JASPAR2024.rds")
}
}
