% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_TGs.R
\name{identify_TGs}
\alias{identify_TGs}
\title{Identify Target Genes via Chromatin Accessibility and Annotation}
\usage{
identify_TGs(
  genome_for_anno = TxDb.Hsapiens.UCSC.hg38.knownGene,
  promoter_range = 50,
  scRNAseq = scRNAseq,
  scATACseq = scATACseq,
  annoDb = "org.Hs.eg.db"
)
}
\arguments{
\item{genome_for_anno}{A TxDb object for genome annotation (e.g., TxDb.Hsapiens.UCSC.hg38.knownGene, TxDb.Hsapiens.UCSC.hg19.knownGene).
Default is TxDb.Hsapiens.UCSC.hg38.knownGene.}

\item{promoter_range}{A positive numeric object. Defines the promoter region around TSS (in base pairs). Default is 50 bp.}

\item{scRNAseq}{A Seurat object containing scRNA-seq data (the initially loaded data) and gene names as rownames.
The \code{scRNAseq@meta.data} contains the seurat_annotation column, which describes cell labels.}

\item{scATACseq}{A Seurat object containing scATAC-seq data (the initially loaded data) with files _fragments.tsv.gz.tbi and _fragments.tsv.gz (in Signac framework).
The \code{scATACseq@meta.data} contains the seurat_annotation column, which describes cell labels.}

\item{annoDb}{"org.Hs.eg.db". Specifies the organism-level annotation package.}
}
\value{
A list containing two elements:
peak_anno_promoter: A data frame of annotated peaks.
TGs: A character vector of unique target gene symbols identified.
}
\description{
This function performs genomic annotation on chromatin accessibility peaks (from scATAC-seq) to identify potential target genes (TGs) with open promoter.
It uses the ChIPseeker package for peak annotation
and generates output files including annotation plots (peak_anno.png), annotation results (peak_anno.csv), and identified TGs (TGs.csv).
}
\note{
The scRNA-seq data and scATAC-seq data need to be loaded as objects named scRNAseq and scATACseq,
respectively, and they must come from the same cells with identical cell labels.
}
\examples{
\dontrun{
library(TxDb.Hsapiens.UCSC.hg38.knownGene)
# Load scRNA-seq data and scATAC-seq data as object scRNAseq and scATACseq with the same cells and cell labels.
# scRNAseq = SeuratData::LoadData("pbmcMultiome", "pbmc.rna")
# scATACseq = SeuratData::LoadData("pbmcMultiome", "pbmc.atac")
scRNAseq = base::readRDS("./scRNAseq.rds")
scATACseq = base::readRDS("./scATACseq.rds")
promoter_range = 50
results_identify_TGs = identify_TGs(
  genome_for_anno = TxDb.Hsapiens.UCSC.hg38.knownGene,
  promoter_range = promoter_range,
  scRNAseq = scRNAseq,
  scATACseq = scATACseq,
  annoDb = "org.Hs.eg.db"
)
base::saveRDS(results_identify_TGs, file = "./1.1 Data Preprocessing - Identify TGs By Annotation/results_identify_TGs.rds")
}
}
