% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_cal.R
\name{R_cal}
\alias{R_cal}
\title{The function R in the evolutionary equations of the paper}
\usage{
R_cal(theta_TF_TG, tao, r1 = 200, r2, r3 = 1, r4 = 1000, r5 = 1)
}
\arguments{
\item{theta_TF_TG}{Numeric (vector). Regulatory strength of transcription factor TF on target gene TG}

\item{tao}{Numeric from 0 to 1. Regulatory threshold τ.
If regulatory strength θ > τ, a regulatory relationship is confirmed (positive sample); o
therwise, it is considered a negative sample.}

\item{r1}{Numeric greater than 0. A hyperparameter. Generally greater than 100. Default is 200.}

\item{r2}{Numeric greater than 0. A hyperparameter.}

\item{r3}{Numeric greater than 0. A hyperparameter. Default is 1.}

\item{r4}{Numeric greater than 0. A hyperparameter. Generally greater than 100. Default is 1000.}

\item{r5}{Numeric greater than 0. A hyperparameter. Default is 1.}
}
\value{
Numeric (vector) from 0 to 1. The computation results of function R.
}
\description{
The function R in the evolutionary equations of the paper
}
\examples{
\dontrun{
R = R_cal(
  theta_TF_TG = c(0.18, 0.23),
  tao = 0.323,
  r1 = 200,
  r2 = 1,
  r3 = 1,
  r4 = 1000,
  r5 = 1
)
}
}
