% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_cal.R
\name{S_cal}
\alias{S_cal}
\title{The function S1 or S2 in the evolutionary equations of the paper}
\usage{
S_cal(s, U, U_tilde)
}
\arguments{
\item{s}{Numeric greater than 0. cGenerally range from 0.2 to 5.}

\item{U}{Numeric ranging from 0 to 1. An intermediate variable.}

\item{U_tilde}{Numeric ranging from 0 to 1. Numeric ranging from 0 to 1.}
}
\value{
Numeric. The computation results of function  S1 or S2.
}
\description{
The function S1 or S2 in the evolutionary equations of the paper
}
\examples{
\dontrun{
print(S_cal(s = s1, U = U1[prev_idx], U_tilde = U1_tilde[prev_idx]))
}
}
