% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_init_params.R
\name{set_init_params}
\alias{set_init_params}
\title{Initialize Model Parameters for Gene Regulatory Network Prediction}
\usage{
set_init_params(
  interest_cell_type_branch_training_set = interest_cell_type_branch_training_set
)
}
\arguments{
\item{interest_cell_type_branch_training_set}{The output of function select_training_set.}
}
\value{
A nested list structure containing initialized parameters for each cell type and branch.
}
\description{
This function initializes the parameter structures for the gene regulatory network prediction model scDEDS.
}
\details{
The function performs the following operations:
\enumerate{
\item Creates a hierarchical parameter structure for each cell type, branch, and regulatory interaction (TF-TG pair).
\item Initializes parameters for the scDEDS model, including:
\itemize{
\item \code{r}: Regulatory interaction parameters (r2_TG, r2_TF) with initial values, upper, and lower bounds.
\item \code{alpha}: Cell group state association parameters (alpha1, alpha2 for TG and TF) for each time point K.
\item \code{beta}: Linear coefficients (beta1, beta2, beta3 for TG and TF) for each time point K.
\item \code{s}: Scaling factors (s1_TG, s1_TF, s2_TG, s2_TF).
\item \code{u}: Non-linear transformation parameters (u11_TG, u21_TG, u311_TG, u321_TG, u331_TG, u12_TG, u22_TG, u312_TG, u322_TG, u332_TG).
\item \code{E~}: Baseline expression levels (E~TG_K-1, E~TF_K-1) for each time point K, initialized as the mean of TGE and TFE from the training set, respectively.
\item \code{v}: Additional scaling parameters (v1_TG, v2_TG, v3_TG).
}
\item For each TF-TG pair, the function uses the mean TFE and TGE values from the training set to initialize the E~ parameters.
\item All parameters are assigned initial values, upper bounds, and lower bounds to constrain the optimization process.
\item The function organizes the parameters in a nested list structure, first by cell type, then by branch, and finally by regulatory interaction (TF-TG pair).
}
}
\examples{
\dontrun{
interest_cell_type_branch_training_set = base::readRDS("./4.1 Build Prediction Model - Select Training Set/interest_cell_type_branch_training_set.rds")
interest_cell_type_branch_init_params = set_init_params(
  interest_cell_type_branch_training_set = interest_cell_type_branch_training_set
)
base::saveRDS(interest_cell_type_branch_init_params, file = "./4.2 BUild Prediction Model/interest_cell_type_branch_init_params.rds")
}
}
